added missing pakages, plus javascript code was sorted on modules
