+++ README.md
@@ -2,6 +2,6 @@
 # ссылка на мой репозиторий https://github.com/DmJavaScript/project11
 Проектная работа 11
 
-Актуальная версия программы v.0.0.1 от 24 октября 2019г.
+Актуальная версия программы v.0.0.2 от 25 октября 2019г.
 Данная версия содержит код проектной работы №9 для проверки правильного размещения проекта на github. 
-Залит проект с общей структурой кода, но пока javascript код не переработан в модульный формат.
+Добавлены необходимые пакеты для обработки кода, javascript код переработан в модульный формат.
