+++ package.json
@@ -4,10 +4,9 @@
   "version": "0.0.1",
   "main": "index.js",
   "scripts": {
-    "build": "NODE_ENV=production rimraf dist && webpack --mode production",
-    "dev": "NODE_ENV=development webpack-dev-server --mode development --open --watch",
-    "dev:server": "webpack-dev-server --config webpack.dev.js",
-    "deploy": "NODE_ENV=production gh-pages -d dist"
+    "build": "cross-env NODE_ENV=production rimraf dist && webpack --mode production --config webpack.prod.js",
+    "dev": "cross-env NODE_ENV=development webpack-dev-server --mode development --open --watch --config webpack.dev.js",
+    "deploy": "cross-env NODE_ENV=production gh-pages -d dist"
   },
   "author": "DmJavaScript",
   "license": "ISC",
@@ -17,17 +16,23 @@
     "@babel/preset-env": "^7.6.3",
     "autoprefixer": "^9.6.5",
     "babel-loader": "^8.0.6",
+    "cross-env": "^6.0.3",
     "css-loader": "^3.2.0",
     "cssnano": "^4.1.10",
+    "file-loader": "^4.2.0",
     "gh-pages": "~2.0.1",
     "html-webpack-plugin": "^3.2.0",
     "mini-css-extract-plugin": "^0.8.0",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
     "postcss-loader": "^3.0.0",
+    "style-loader": "^1.0.0",
+    "tapable": "^1.1.3",
     "uglifyjs-webpack-plugin": "^2.2.0",
     "webpack": "^4.41.1",
     "webpack-cli": "^3.3.9",
-    "webpack-md5-hash": "0.0.6"
+    "webpack-dev-server": "^3.9.0",
+    "webpack-md5-hash": "0.0.6",
+    "webpack-merge": "^4.2.2"
   },
   "dependencies": {
     "build": "^0.1.4",
